+++ src/pydantic_market_data/interfaces.py
@@ -1,6 +1,14 @@
-from typing import List, Optional, Protocol
+from datetime import date
+from typing import Protocol
 
-from .models import History, HistoryPeriod, SecurityCriteria, Symbol
+from .models import (
+    History,
+    HistoryPeriod,
+    PriceInput,
+    SecurityCriteria,
+    Symbol,
+    TickerInput,
+)
 
 
 class DataSource(Protocol):
@@ -8,20 +16,26 @@ class DataSource(Protocol):
     Interface for a financial data source.
     """
 
-    def search(self, query: str) -> List[Symbol]:
+    def search(self, query: str) -> list[Symbol]:
         """
         Search for security by ISIN, symbol, or name
         """
         ...
 
-    def resolve(self, criteria: SecurityCriteria) -> Optional[Symbol]:
+    def resolve(self, criteria: SecurityCriteria) -> Symbol | None:
         """
         Resolve security based on provided criteria
         """
         ...
 
-    def history(self, ticker: str, period: HistoryPeriod = HistoryPeriod.MO1) -> History:
+    def history(self, ticker: TickerInput, period: HistoryPeriod = HistoryPeriod.MO1) -> History:
         """
         Fetch historical data for a ticker
+        """
+        ...
+
+    def validate(self, ticker: TickerInput, target_date: date, target_price: PriceInput) -> bool:
+        """
+        Validates if the ticker traded near the target price on the target date.
         """
         ...